Refactor namespaces for weather model
