+++ app/Models/Weather/Records/Wind.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App;
+namespace App\Models\Weather\Records;
 
 use App\Weather\Record;
+use Illuminate\Http\Request;
 
 class Wind extends Record
 {
